% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alt_denominator.R
\name{calc_alt_denominator_monthly}
\alias{calc_alt_denominator_monthly}
\title{Calculate Alternative denominator}
\usage{
calc_alt_denominator_monthly(ind_value, denominator)
}
\arguments{
\item{string}{a dataframe from sormas}
}
\value{
a dataframe.
}
\description{
Calculate Alternative denominator
}
\examples{
cleaned_measles_data <- "sormas_cleaner(file = "path")"

}
