% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alt_denominator.R
\name{get_2_dose_alt_denominator}
\alias{get_2_dose_alt_denominator}
\title{Get alternative denominator for 2 dose antigen e.g Measles, Penta}
\usage{
get_2_dose_alt_denominator(
  sormas_cleaned,
  first_dose,
  later_dose,
  first_dose_given,
  later_dose_given,
  denominator_data
)
}
\arguments{
\item{first_dose}{the earlier dose coverage e.g Measles 1 coverage}

\item{later_dose}{the later dose coverage e.g Measles 2 coverage}

\item{first_dose_given}{the earlier dose given e.g Measles 1 given}

\item{later_dose_given}{the later dose given e.g Measles 2 given}

\item{denominator_data}{alternative denominator data}

\item{string}{a dataframe from sormas}
}
\value{
a dataframe.
}
\description{
Get alternative denominator for 2 dose antigen e.g Measles, Penta
}
\examples{
cleaned_2_dose_data <- "sormas_cleaner(file = "path")"
}
