% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_cases_alt_denominator.R
\name{get_alt_denominator}
\alias{get_alt_denominator}
\title{Get alternative denominator for a disease of your choice}
\usage{
get_alt_denominator(coverage, sormas_cleaned, doses_given, denominator_data)
}
\arguments{
\item{string}{a dataframe from sormas}
}
\value{
a dataframe.
}
\description{
Get alternative denominator for a disease of your choice
}
\examples{
cleaned_measles_data <- "sormas_cleaner(file = "path")"
}
