% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcv_different_sources.R
\name{get_mcv_different_sources}
\alias{get_mcv_different_sources}
\title{add measles}
\usage{
get_mcv_different_sources(mcv1_annualized_all_time, mcv2_annualized_all_time)
}
\description{
add measles
}
