% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_cases_alt_denominator.R
\name{get_measles_alt_denominator}
\alias{get_measles_alt_denominator}
\title{Get alternative denominator for measles only}
\usage{
get_measles_alt_denominator(
  sormas_cleaned,
  mcv1,
  mcv2,
  measles1_given,
  measles2_given,
  denominator_data
)
}
\arguments{
\item{string}{a dataframe from sormas}
}
\value{
a dataframe.
}
\description{
Get alternative denominator for measles only
}
\examples{
cleaned_measles_data <- "sormas_cleaner(file = "path")"
}
