% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stock_analysis.R
\name{get_stock_analysis}
\alias{get_stock_analysis}
\title{getting other diseases with one dose}
\usage{
get_stock_analysis(received, opening_balance, opened_used, doses_given)
}
\description{
getting other diseases with one dose
}
