% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sormas_cleaner.R
\name{sormas_cleaner}
\alias{sormas_cleaner}
\title{Clean sormas data}
\usage{
sormas_cleaner(
  file,
  skip = 2,
  disease = c("diphtheria", "measles", "yellow fever", "meningitis (csm)")
)
}
\arguments{
\item{string}{a dataframe from sormas}
}
\value{
a dataframe.
}
\description{
Clean sormas data
}
\examples{
cleaned_measles_data <- "sormas_cleaner(file = "path")"
}
